[viewer] add shortcuts about view prev/next image
When the input_box is opened, the left/right shortcuts of viewing
prev/next image need to be unbined, since these two keys are used for
moving the cursor when inputting in the input_box.

As an alternative, the shortcuts of up, down, mouse wheel up and down
are used for viewing prev/next image. Those shortcuts won't be affected
by the input_box.