[dict_editor] fix bug in export_dict
This function decides the source of exporting the dictionary based on
whether it is saved and the user's choice. If it is saved, it will be
exported from the saved OrderedDict, otherwise it will be exported from
the current Treeview.
Therefore, if the value in the dictionary is a list, it needs to be
merged into a string, separated by spaces.